Relnotes: Support projects without deps.bzl file
The WORKSPACE stanza of the release notes should not always contain instructions to load a deps.bzl file, since projects may not have such a file. This behavior can be controlled by setting the has_deps_file parameter of the print_rel_notes macro.
It defaults to True for backwards compatibility.