Add pkgfilegroup for archive-independent file staging
Currently, package contents need to be assembled in each of their `pkg_*` rules,
and each one has a different mechanism for doing so.  This provides a unified
interface, the `pkgfilegroup` rule, for accomplishing this task. Packaging rules
can be written to take advantage of this in future changes.

Analysis tests are also included.

This is inspired by #36.

This is currently incomplete.  Features in development include:

- File renaming
- Non-file creation (e.g. directories, device nodes)