Add pkgfilegroup for package-independent destination mappings (#128)
* Add pkgfilegroup for package-independent destination mappings

This adds an experimental rule, `pkgfilegroup`, along with associated Providers,
that gives rule developers and users a consistent mechanism for using the output
of bazel targets in packaging rules.

Inspired by #36.

Other capabilities that are provided by this that were not mentioned in #36 are:

- Creation of empty directories (`pkg_mkdirs`)
- Exclusion of files from a `pkgfilegroup` (`excludes`)
- Renames of files in a `pkgfilegroup` (`renames`)

* Add analysis tests for pkgfilegroup and friends

This provides some analysis tests for various features of `pkgfilegroup` and
`pkg_mkdirs`.  See #128.

You can run them by invoking `bazel test experimental/...` from the `pkg`
directory

This implementation of pkgfilegroup was inspired by #36.