Provide tests for experimental/pkg_rpm.bzl
This change provides a basic test system for testing changes to the experimental
`pkg_filegroup`-based RPM generator.  The test framework itself is a python script
that makes several calls out to `rpm(8)`'s query command on a generated RPM, and
comparing output information to expected values.

This could also be theoretically applied to the existing RPM generator, if so
desired.

There is more that could be tested here, but this is adequate in the meantime to
demonstrate how this might be applied to other RPM features.