Ensure the experimental/ tree lands in the output release archive
The `experimental/` tree containing the then `pkgfilegroup` rule was missing
from the 0.2.5 release archive.

This change ensures that the files in the experimental/ tree are included in
release packages so users can use the rules within.

Testing was done by trying to run the experimental `pkg_rpm` rule on a dummy
project importing rules_pkg.