Provide pkg_mklinks for creation of in-package symbolic links
Symbolic links are often included within packages, referring to files and
directories within and without.

This commit provides a rule, `pkg_mklinks`, which allows for the creation of
arbitrary symbolic links within a package, and support within the experimental
`pkg_rpm` rule to emit them.

`buildifier` was also opportunistically applied to files in this change.