Mass `pkgfilegroup` -> `pkg_filegroup` rename
This change renames all instances of `pkgfilegroup` to `pkg_filegroup`, and
renames `experimental/genpkg.bzl` to `experimental/pkg_filegroup.bzl`.

Names of tests and documentation have been updated accordingly.  The change was
mostly made mechanically using `sed(1)`, with minor cleanup afterward.

This change makes the names within the experimental/ directory consistent, and was
requested in the review of #129.