Change `gen_rpm` back to `pkg_rpm` in experimental/rpm.bzl (#159)
The previous change (2536b7b) reverted some of the naming changes in
experimental/rpm.bzl and elsewhere in the codebase.  This was due to a
cherry-pick that looks otherwise correct.

This change adjusts `gen_rpm` back to `pkg_rpm`, and `genrpm` to `rpm` in
various files in experimental/.

The experimental/ tree was checked for the presence of `gen_rpm` and `genrpm`
with `grep(1)`; nothing was found.