Add support to generate stardoc.
This is done in a manner so there is no new runtime dependency on bazel-skylib. The bzl_library needed as input to stardoc is only created within the distro directory, which is not part of the released package.

Only a trivial cleanup of the docs within the rules was done in this change.