Fix `pkg_rpm` `source_date_epoch` attribute
The existing `source_date_epoch` attribute for `pkg_rpm` is passed in as
an `int`, and is then appended to a string.  This, naturally, fails.
The `int` needs to be stringified first.

Additionally, since the [SOURCE_DATE_EPOCH] value can be any valid UNIX
epoch value, account for the possibility of it being set to 0 in the
Starlark code.

[SOURCE_DATE_EPOCH]: https://reproducible-builds.org/specs/source-date-epoch/#idm55