Allow experimental RPM builder tests to pass on CentOS 7 and CentOS 8
Tests on CentOS 8 with the RPM generator test code fail due to:

- RPM file changing magic
- Directory size calculation changes
  (https://github.com/rpm-software-management/rpm/commit/2cf7096ba534b065feb038306c792784458ac9c7)

This change accounts for these issues by ignoring RPM spec file "magic" and
throwing out the whole concept of "size" to begin with (hashes are good enough),
respectively.

Quoting was also opportunistically adjusted in the manifest CSV `genrule`.