Provide tests for experimental/pkg_rpm.bzl (#160)
* Provide tests for experimental/pkg_rpm.bzl

This change provides a basic test system for testing changes to the experimental
`pkg_filegroup`-based RPM generator.  The test framework itself is a python script
that makes several calls out to `rpm(8)`'s query command on a generated RPM, and
comparing output information to expected values.

This could also be theoretically applied to the existing RPM generator, if so
desired.

There is more that could be tested here, but this is adequate in the meantime to
demonstrate how this might be applied to other RPM features.

* Switch up python string concat

* Improve test manifest formatting, test completeness

- Fix file size calculations.  Turns out the issue was related to following
  symlinks.

- Take @aiuto's suggestion for formatting the RPM CSV manifest writing for
  improved maintainability

* Allow experimental RPM builder tests to pass on CentOS 7 and CentOS 8

Tests on CentOS 8 with the RPM generator test code fail due to:

- RPM file changing magic
- Directory size calculation changes
  (https://github.com/rpm-software-management/rpm/commit/2cf7096ba534b065feb038306c792784458ac9c7)

This change accounts for these issues by ignoring RPM spec file "magic" and
throwing out the whole concept of "size" to begin with (hashes are good enough),
respectively.

Quoting was also opportunistically adjusted in the manifest CSV `genrule`.

* Add centos 7 test target; tag rpm tests as not runnable on macOS or Windows

* Disable the experimental RPM packaging test until CI has rpmbuild(8)

See #183.

* Whoops, forgot macOS.

Add some more explanatory comments while I'm at it.