Provide broken test for bug in merged tars with big file names
When adding tarballs as input for other tarballs (i.e. `pkg_tar` that
depends on another `pkg_tar`), every file with a name that exceeds 98
characters will not land in the new root directory (`package_dir`
attribute), but rather on the actual root.

Consider the following example:

    pkg_tar(
        name = "foo",
        package_dir = "/foo",
        strip_prefix = ".",
        srcs = [ ... ],
        deps = [":big-names"],
    )

    pkg_tar(
        name = "big-names",
        strip_prefix = ".",
        srcs = [
            "ninety-eight-characters-seems-to-be-the-maximum-limit-for-merged-tarballs-in-the-rules_pkg-for-bzl",
            "ninety-eight-characters-seems-to-be-the-maximum-limit-for-merged-tarballs-in-the-rules_pkg-for-bzl-",
        ],
    )

The foo tarball will have the following structure:

    .
    ├── foo
    │   └── ninety-eight-characters-seems-to-be-the-maximum-limit-for-merged-tarballs-in-the-rules_pkg-for-bzl
    └── ninety-eight-characters-seems-to-be-the-maximum-limit-for-merged-tarballs-in-the-rules_pkg-for-bzl-

Filenames with 98 characters or less get included in the right
directory, but the larger ones go to the root. A deep directory
structure behaves the same way:

    .
    ├── foo
    │   └── ninety/eight/characters/seems/to/be/the/maximum/limit/for/merged/tarballs/in/the/rules_pkg/for/bzl
    └── ninety/eight/characters/seems/to/be/the/maximum/limit/for/merged/tarballs/in/the/rules_pkg/for/bzl-

And the size of the package_dir has no influence: "/f", "/foo" or
"/foo/bar/baz/quux" behaves exactly the same way.