Explicitly use GNU_FORMAT instead of PAX_FORMAT for tars
Up until Python 3.8, the default format used to be GNU_FORMAT. On 3.8,
it changed to PAX_FORMAT, which is supposed to also support big file
names. There seems to be a bug in it, exposed by the test in my previous
commit.

This commit ensures that even with Python >= 3.8 long file names still
work, by explicitly requesting GNU_FORMAT when writing.