Mockup of hierarchical pkg_filegroup implementation
This change is intended to start a more detailed discussion on how we might make
a "nestable" `pkg_filegroup` rule.

The mechanism for calling rules out in other rules was tested and is known to
work, but I wonder about what to do with the `name` attribute.  A thought that
comes to mind is that `pkg_file` could emit a specification that a (currently
nonexistent) `pkg_filegroup` rule could use to provide a name to a "real"
`pkg_file` rule (or however it is implemented).