Provide a more concise package "manifest" specification
---

In using `pkg_filegroup` in our codebases, we've found that it takes a
significant amount of vertical space for complex packages.  For simple
mappings (e.g. copy a target/file to a destination), we've found that a simple
tabular format provides a succinct way to specify contents without sacrificing
much in the way of readability.

At this time, I'm looking to start a discussion on the idea presented here.  The
idea itself is essentially agnostic to the implementation of `pkg_filegroup`,
but `pkg_filegroup` needs some changes in order to make the usage below more
idiomatic: there's currently no means to consolidate multiple differing
`pkg_filegroup`-like rules into a single rule.  A potential idea for this is
discussed in #212.

---

This change provides a simple csv-like mechanism for specifying package
contents.  Instead of specifying multiple `pkg_filegroup`s, you can instead say
something like:

```python
manifest = [
    ("/usr/bin/foo", "copy", "unix=0755", ":foo"),
    ("/etc/foo.d",   "mkdir" "unix=0755", ""),
    ...
]

manifest_rules = pkg_list_manifest(
    name = "manifest",
    manifest = manifest,
    default_attrs = "user=root;group=root",
)

pkg_rpm(
    name = "my-rpm",
    data = manifest_rules + [
        # Complex pkg_filegroup rules here
    ],
    ...
)
```

Which, for simple install-and-create operations, is enough and is much easier to
read.

See the in-file documentation in `pkg/experimental/manifest/manifest.bzl` for
more details on how the manifest is intended to be structured.

Tests were also provided, runnable as `//experimental/manifest:manifest_test`.
They are currently incomplete, but are enough to prove basic correctness so far.

The one thing that is not yet clear to us is how we should specify and order the
manifests.  Some potential examples are provided in
`pkg/experimental/manifest/examples`, with further information described in
`pkg/experimental/manifest/examples/README.md`.