experimental/rpm.bzl: Make compression settings configurable
This change adds a new attribute `binary_payload_compression` to the
experimental `pkg_rpm` rule.  It is implemented in terms of command-line options
to `rpmbuild(8)`, which are passed in via `make_rpm.py`.

To allow for users to provide a template-level default, we cannot use the
provided `_binary_payload` macro.  We instead need to provide a custom one, and
then conditionally use it as `_binary_payload` instead.  The existing RPM spec
templates were updated to account for this.  Tests were also provided.

This can also be added to the non-experimental `pkg_rpm` rule in the future; the
code would be similar.

Fixes #239.