experimental/rpm.bzl: Make compression settings configurable (#240)
This change adds a new attribute binary_payload_compression to the
experimental pkg_rpm rule. It is implemented in terms of command-line options
to rpmbuild(8), which are passed in via make_rpm.py.

We do this by instructing rpmbuild to define the _binary_payload macro on its command line. Tests are also provided.

This can also be added to the non-experimental pkg_rpm rule in the future; the
code would be similar.

Fixes #239.