Provide providers for new packaging rule interfaces
This change provides a set of three providers that can be used to represent
actions, and one that can represent a group of related actions.

These are intended to be used in rules like the `pkg_filegroup` interfaces used
in pkg/experimental, but with more flexibility and standardization.

`buildifier` was also applied opportunistically.