Create rules for emitting PackageFilesInfo, PackageDirsInfo
This change creates rules for emitting `PackageFilesInfo` (`pkg_files`) and
`PackageDirsInfo` (`pkg_mkdirs`) providers.

The implementations and tests of these rules are largely similar to the rules
that currently exist in `experimental/pkg_filegroup.bzl`, with a few notable
differences:

- `pkg_filegroup` is replaced with `pkg_files`.  `pkg_filegroup` will be
  re-implemented as a grouping provider in a future change.

- `pkg_files` now checks for files that map to the same destination.  The check
  was previously deferred to the packaging rules themselves.  This behavior is
  now tested.

- `attr` checks for `pkg_files` and friends are now no longer strict.  Only the
  "unix" attribute is verified.  The logic for how to check this isn't yet
  complete, but it's enough to make forward progress.

- `section` for `pkg_files` and `pkg_mkdirs` is removed.  It will be replaced
  with more specific attributes as needed.

- `make_strip_prefix` is removed in favor of the `strip_prefix` pseudo-module.

`pkg_mkdirs` is effectively unchanged, other than differences in the providers.
The tests are also effectively unchanged, excepting those regarding `section`
and the ones added above.