Refactior usage of xz compression so that support for it is limited to one .bzl and one .py file.
Context: #275

Some Python implementations and/or build environments do not have support for xz compression.  In order to make it easier to developer and use the package on those platforms pkg.bzl now provides an indicator of support. That is used in the test cases to conditionally add tests for xz compressed tar files. The python code follows a similar pattern.

This allows someone importing the rules to change only one or two locations to change the test patterns to conform to the constriants of their environment.

Next step: Make the connection between the Starlark and Python code automatic.