Provide "official" pkg_filegroup-using RPM rule
This change provides a functional instance of a `pkg_filegroup`-using packaging
rule, namely `pkg_rpm`.  This is derived from the rule defined in
`pkg/experimental/rpm.bzl`, with notable differences being:

- Usage of new providers and different paths.

- Restructuring of how the `%files` metadata is built out.  It was restructured
  for clarity, and now makes use of the new `rpm_filetag` mapping attribute
  instead of `section`, which was since removed.

- Inputs to `pkg_rpm` are now specified with the `srcs` attribute instead of
  `data`, to be consistent between runtime requirements and compile-time sources.

- Conflicting content error detection and emission is now streamlined.

Tests are available, but are not currently included as a part of this change.  I
will post them as a separate change initially,

The location specified here is not going to be where these end up; we should
know where early next week (2021-03-01).