Provide pkg_filegroup-using pkg_rpm contents/analysis tests
This change provides a number of unit tests for the pkg_filegroup-using
`pkg_rpm` rule, meant to ensure that packages are formed as expected and
conflict detection is properly implemented.

The main differences between these and the tests in experimental/tests are:

- Analysis tests for conflict detection
- Fixed bugs WRT contents detection
- Path changes

This is likely not the final destination for these test cases, and they be
combined with the rules implementation, depending on the size of the diff.  See
#303 for more details.