Allow for more structured package mapping rule attributes
This change modifies the `attributes` attribute of `pkg_files` and `pkg_mkdirs`
to allow for more structured inputs than a list of strings.

For the lack of a better option, the solution for this right now is to serialize
the data in terms of JSON strings, which will be assembled using the new
`pkg_attributes` macro.  Users should always use `pkg_attributes` for this
purpose.

The JSON is decoded within the mapping rules and passed out through their
providers.  Their documentation was updated where appropriate.

A better option would be to pass Providers full of structured data directly into
rules, but, unfortunately Bazel does not support this at this time.