Remove old pkg_filegroup rule and tests; cleanup (#316)
The old `pkg_filegroup` ruleset in `pkg/experimental/pkg_filegroup.bzl` is
deprecated in favor of the one in `pkg/mappings.bzl`.  Now that no other rules
use it, we can remove the rules and their tests.

This change does that, and additionally cleans up some excess targets in
`pkg/experimental/BUILD`.