experimental/rpm.bzl: support filename templating; remove "outputs" rule attribute
This change modifies the `pkg_filegroup`-using `pkg_rpm` rule to use the same
naming attributes and output system as other packaging rules, specifically:

- `package_variables` corresponds to a rule providing `PackageVariablesInfo`,
   used for templating of various package attributes.

- `package_file_name` corresponds to the output filename, and uses the value
  provided to `package_variables` to construct an "expanded" output file name.

- `package_name` corresponds to the "name" of the package, as recognized by the
  package manager.  This replaces `rpm_name`.

- The expanded output file name is identified via a `PackageArtifactInfo`
  provider.  The default, unexpanded name is derived from `package_name`, or
  `name` if not provided.

If `package_file_name` is not provided, then the default name is formatted
according to the preferred RPM "NVRA" format.  File names will look like
`$name-$version-$release.$architecture.rpm`.

As a side-effect, this removes the deprecated `outputs` output from `pkg_rpm`.
This resolves #282 for this rule, but not the soon-to-be-legacy rule in
`rpm.bzl`.   If it is desired to do this, I will do so in a follow-on-change.

Since (both versions of) `pkg_rpm` are structured differently from the other
packaging rules, `setup_output_files` in `private/util.bzl` was modified to
support deriving the default output file name from a macro argument instead of
relying on `ctx.outputs.out`, as it is not present in `pkg_rpm`.  The
documentation was also updated to mention this and to improve clarity.

Documentation was also opportunistically fixed in `experimental/rpm.bzl`.

Tests are also provided.