Support long file names by ignoring 'path' PAX header. (#250)
The 'path' PAX header is normally used to handle long file names. When re-packaging a tar
containing a file with a long name we wouldn't touch that header and the file wouldn't
go to its new path. To fix that we simply delete the header before writing as the tarfile
library will handle writing a new one if needed.

Side note: Right now this code is not following the 4-spaces styleguide. I would've fixed
it now but then the change I'm making would be hidden in the noise.

See also #216.