Provide `filter_directory` for TreeArtifact processing
TreeArtifacts (directory outputs) are effectively a black box to Bazel rules.
You can pass them to commands easily, but Bazel cannot operate directly on
them: making the `pkg_filegroup` framework impossible to use on them.

While avoiding TreeArtifacts is preferred in some cases, in many other cases, it
is simply unavoidable, but you still need to do some processing on them to get
the directory structure exactly right.  A common solution to this problem is to
use something like a `genrule` or `pkg_zip` to change the directory structure
and produce an archive that contains the desired structure, and then to extract
it again to get the "proper" TreeArtifact.

This is somewhat wasteful.  To alleviate this issue, this change provides a
`filter_directory` rule which provides a small set of operations that might be
useful to process TreeArtifacts.  These are implemented in the style of
`pkg_files`, and is deliberately primitive.  Additional operations (including
arbitrary path rewriting) can be implemented at a later time if desired.

Extensive (but perhaps incomplete) tests are also provided.