Allow runfiles to be used alongside the `rpmbuild` toolchain (#329)
The current rpmbuild toolchain setup do not seem to propagate the runfiles
needed by `rpmbuild` when `pkg_rpm` calls `make_rpm.py`, despite it apparently
having everything it needs passed into `ctx.actions.run#tools`.

It seems as though passing in the associated `FilesToRunProvider` fixes this,
although I'm not sure why.  This change is implemented here, as well as pulling
accessing the fields of `DefaultInfo` via the provider rather than directly
through the `Target` structure (marked as [deprecated] in most recent
documentation).

Fix was implemented in both instances of `pkg_rpm`; the code is sufficiently
similar to allow for copy+paste to be used effectively.

[deprecated]: https://docs.bazel.build/versions/master/skylark/lib/Target.html

Fixes #328.