Improve pkg_filegroup-using pkg_rpm TreeArtifact installation (#319)
`pkg_rpm` with `pkg_filegroup` never properly consumed TreeArtifacts (directory
outputs).  This is due to two factors:

- In RPM packages, adding a directory as a file results in the directory being
  owned and every file under it having consistent permissions.  In our
  experience, this has led to confusing behavior and can make RPM package
  content listings harder to understand.

- The command for copying files into place does the wrong thing when the target
  directory already exists for any reason.

This change remedies the above by enumerating TreeArtifact directory structures
during the execution phase (as they are not knowable at analysis time), and
passing the outputs to a command that creates additions to the install scripts
and `%files` manifests.  Files in TreeArtifacts are now installed and owned by
the RPM individually.  Attributes are applied identically to all files in the
TreeArtifact, and TreeArtifact processing is only performed if any are detected.

It is currently not possible to any additional processing to TreeArtifacts,
including excluding files and stripping prefixes (#269).  Until such time,
workarounds such as using `genrule`s or other custom rules should be used.

Regression tests are also provided.

Additional changes include:

- Fixing some uninclusive language.

- Refactoring the "RPM contents manifest" code for the `pkg_rpm` into a library.

- Opportunistically applying `buildifier`.

- Removing an unused test symlink.

NOTE 1: Other possible implementations of this involve changing the existing
`PackageFilesInfo` provider and/or creating a new one that represents
TreeArtifacts.  This was not explored, but many of the implementation details
used here can be re-purposed should that be considered.

NOTE 2: The change does not attempt to do conflict checks between TreeArtifact
installations and other TreeArtifact installations, or
files/directories/symlinks already identified at analysis time.  To reduce the
complexity of this otherwise large change, this feature currently deferred.  My
current (untested) understanding is that if this happens, duplicates in `%files`
will be detected, and cause most versions of `rpmbuild` to fail.

Fixes #292.