Codify how to remove TreeArtifact base directories in `pkg_files`  (#330)
We often have TreeArtifacts (directory artifacts) that are built to be installed
directly onto the filesystem in packages, but they sometimes need to be renamed.
As of #319, there are two ways to accomplish this:

1) Rename the TreeArtifact to the destination
2) Rename the TreeArtifact to nothing

I believe that (2) is clearer, as the `prefix` provided to `pkg_files` stands
out more than an entry in `renames`.  This change codifies a workaround we have
tried out internally -- renaming to "nothing".  While "nothing" itself could be
a bit misleading, we can instead represent the desire for this via the constant
`REMOVE_BASE_DIRECTORY`, represented as a non-printable string. (1) is still
allowed.

This change does not attempt to do anything related to updating paths inside
TreeArtifacts.  This will be attempted in a follow-on change.

Tests and documentation updates are also provided.

This change also moves the common analysis test boilerplate to its own location
in `pkg/tests/util/defs.bzl`, where it will be more generally accessible in
future changes.

Partially resolves #269.