Deprecate pkg/rpm.bzl and move it to pkg/legacy/rpm.bzl (#337)
The `pkg_filegroup` framework has evolved considerably since its creation, and
the `rpm` packaging rule that uses it (in `pkg/experimental/rpm.bzl`) has as
well.  We believe it is time to move it out of `experimental` and into general
use.

This change prepares for this by:

- Moving the existing `pkg/rpm.bzl` module and its associated tests to new
  package in `pkg/legacy`.

- Replacing the contents `pkg/rpm.bzl` with a compatibility macro that allows
  for easily switching between the two (the interfaces are significantly
  different).  The macro switches between the two based on mandatory attributes
  that are available in one instance of `pkg_rpm` but not the other.

CI, packaging, and test targets were also updated as needed.

Part of #307.