Provide `filter_directory` for basic TreeArtifact processing (#331)
TreeArtifacts (directory outputs) are effectively a black box to Bazel rules.
You can pass them to commands easily, but Bazel cannot operate directly on them,
making the `pkg_filegroup` framework impossible to use in this case.

While avoiding TreeArtifacts is generally preferred, doing so may be difficult
or simply unavoidable.  When you do need to use them, you often still need to do
some processing to achieve the desired directory structure.  A common solution
to this problem is to use something like a `genrule` or `pkg_zip` to produce an
archive that contains the desired structure, and then to extract it in a rule to
get the desired TreeArtifact.  Archive creation and extraction is necessary
because `genrule`s do not support emitting directories in a reliable manner.

Creating intermediate archives is wasteful.  To help alleviate this issue, this
change provides a `filter_directory` rule which provides a small set of
operations that might be useful to process TreeArtifacts.  These operations are
implemented in the style of `pkg_files`, and are deliberately primitive.
Additional operations (including arbitrary path rewriting) can be implemented at
a later time if desired.

Extensive (but perhaps incomplete) tests are also provided.