Fix mapping tests so they can run on Windows.
1. Eliminate sh_binary
The problem here is that sh_binary creates an extra .exe on Windows.
That is, sh_binary(name="foo",...) manifests an output named "foo.exe".
This breaks tests that used pkg_* to manipulate outputs of a sh_binary
and compare the returned provider data to expected expected results.

I could hack up the tests with windows specific code or a detector for
that case, but that is just a new kind of fragile.  Instead, I added
a fake_artifact rule that creates a text output which would work
as a shell or bat script and simply forwards some inputs out as
providers. It can control if the output is supposed to be executable
or not, so we can use that in future tests.