Change compute_data_path to use ctx as an input instead of output_file.
This is a refactoring before the PR where we introduce pkg_files* support
for pkg_tar & pkg_zip. The intent for that PR is that there will be a
helper method like

```
add_pkg_srcs(ctx, manifest, ctx.attr.srcs)
```

That will call comput_data_path and then use that per file in ctx.attr.srcs
to compute the destination path. Using ctx instead of an arbitrary
output file feels cleaner.

This was more broadly tested in the work in this branch,
https://github.com/aiuto/rules_pkg/tree/tar_pkg_files.
There I have most of the full PR, but pulling this out makes an
easier review.