Provide `pkg_install` for a "make install"-like experience in Bazel
This change provides the `pkg_install` macro, which allows users to create
runnable installation scripts directly from Bazel BUILD files.

Installation scripts are generated from a template that consumes the manifest
files currently known to `pkg_tar`.  They are intended to be run by `bazel run`,
although they can be run within Bazel as tools if needed.  See the documentation
of `pkg_install` in `pkg/install.bzl` for usage details.

`pkg_install` is currently incomplete.  In particular, the implementation so far
only supports installing files and directories on UNIXy platforms.  The
following are not supported:

- Changing of Windows permissions outside of the context of UNIX permissions
- Installation of symbolic links and TreeArtifacts
- Control of default permissions/attributes

The following are not adequately tested due to other complications:
- Changing of file ownership (UNIX group/user)

Full support for Windows is likely to occur in separately.

Other changes include:
- Support for emitting "short" paths as manifest destinations (needed to support
  `bazel run`)
- Manifest functionality has been isolated to a common python
  module (`//private/manifest`) to reduce code duplication.

Advances #308.