Modularize manifest python code
The manifest code used in pkg_tar has uses outside of pkg_tar.  Let's modularize
it to reduce the number of places it has to be duplicated (it needs to be
duplicated in `pkg/private/pkg_files.bzl` to be loaded) into starlark.

As a part of this, we create an entry type stringifier for ManifestEntry, and a
`collections.namedtuple` representing the fields in the entry.

`pkg/private/build_tar.py` was also opportunistically modified to use all of the
above.