manifest: Allow for TreeArtifacts in PackageFilesInfo
The package-independent JSON manifests currently do not consider whether a
"file" in a `PackageFilesInfo` provider is actually a file or a directory.  This
change provides this awareness.

This also fixes up some other (apparent) logic issues related to directories,
and additionally provides a test.  Since the tests now involve generated
TreeArtifacts, this change also forces the manifest generation tests to use the
manifest "short_path".

There may be a better way to do this, but it seems to work for now.