[pkg_deb] fix computation of changes file file name (#418)
* [pkg_deb] fix computation of changes file name

When >1 deb packages are defined with the same name and version, but
different arch (e.g. `amd64` and `arm64`), there are multiple
conflicting `changes` files, because anything after the version is not
taken into account (e.g. `arch`). Excerpt from the attached test case,
but unpatched `pkg.bzl`:

```
ERROR: file 'tests/test_deb_multiarch_1.changes' is generated by these conflicting actions:
```

The patch takes the dots in version into account, and changes the
filename only by the last dot, of the presumed `.deb`.

* Revert test

The test could be better and it should be in a new place now.
So, I'll merge this PR now, and we can work out a better test later.

Co-authored-by: aiuto <aiuto@google.com>