manifest: Allow for TreeArtifacts in PackageFilesInfo (#421)
The package-independent JSON manifests currently do not consider whether a
"file" in a `PackageFilesInfo` provider is actually a file or a directory.  This
change provides this awareness.

This also fixes up some other (apparent) logic issues related to directories,
and additionally provides a test.  Since the tests now involve generated
TreeArtifacts, this change also forces the manifest generation tests to use the
manifest "short_path".

There may be a better way to do this, but it seems to work for now.

Additionally, this change fixes some Windows-specific path separator consistency
issues exposed in `pkg_tar` by the aforementioned change.  Thanks to
@beasleyr-vmw for investigating this.

Co-authored-by: Ryan Beasley <beasleyr@vmware.com>