Consolidate pkg_rpm input processing loop
The original `pkg_filegroup`-using `pkg_rpm` design had separate loops for
conflict detection and contents processing.  This change consolidates the two
loops.  The combined loop is a bit long, but the code around this point needs to
be adjusted for #456 anyway, so the length problem will be dealt with then.

Much of this code can be eliminated when we move pkg_rpm to use the JSON
manifests.

Fixes #191.