Allow pkg_rpm to accept mapping providers directly
`pkg_rpm` initially took a strict stance on the `pkg_filegroup` framework.  In
particular, it would only accept `PackageFilegroupInfo`s, and not any of the
other providers.

It was found that direct access to the mapping
providers (e.g. `PackageFilesInfo`) was highly beneficial.  This change
implements this for `pkg_rpm`.  A large amount of the provider-handling
functionality is now abstracted into functions.

The corresponding error message handling logic was improved, and tests were
provided.

Some reordering was also done by `buildifier`.

Depends on #458.

Resolves #456.