Allow pkg_rpm to accept mapping providers directly (#460)
`pkg_rpm` initially took a strict stance on the `pkg_filegroup` framework.  In
particular, it would only accept `PackageFilegroupInfo`s, and not any of the
other providers.

As `pkg_filegroup` implementation for other rules proceeded, it was found that
direct access to the mapping providers (e.g. `PackageFilesInfo`) was highly
beneficial.  This change implements this for `pkg_rpm`.  A large amount of the
provider-handling functionality is now abstracted into functions.  While they
are redundant with the code in `pkg/private/pkg_files.bzl`, utilizing it
requires nontrivial effort (see #459).  This is an easy workaround until we can
migrate `pkg_rpm` to use it.

The corresponding error message handling logic was also improved, and tests were
provided.

Some reordering was also done by `buildifier`.

Resolves #456.