Fix missing `%dir` RPM filetag when `PackageDirsInfo` is provided (#473)
In #460, I added support for `pkg_rpm` accepting mapping
providers (e.g. `PackageFilesInfo`) instead of just grouping
providers (e.g. `PackageFilegroupInfo`).

In a copy+paste error, I neglected to ensure that the `%dir` tag is passed into
the RPM `%files` manifest when PackageDirsInfo is provided directly to
`pkg_rpm.  Without this, `rpmbuild` will recursively add in the entire directory
mentioned, not just the directory itself (the intent of`PackageDirsInfo`).

This change fixes this problem, and adds a regression test.